Preserve virtual feed tails in constrained viewports
